+++ instrumentation/opentelemetry-instrumentation-arangodb/src/opentelemetry/instrumentation/arangodb/arangodb_attributes.py
@@ -1,0 +1,9 @@
+QUERY_ALLOW_DIRTY_READS = "arangodb.query.allowDirtyReads"
+
+QUERY_CACHE = "arangodb.query.cache"
+
+QUERY_STREAM = "arangodb.query.stream"
+
+RESPONSE_CACHED = "arangodb.response.cached"
+
+RESPONSE_COUNT = "arangodb.response.count"